+++ src/pages/Diagram.jsx
@@ -1,0 +1,5 @@
+import { Box } from "@mui/material";
+
+import NodeLinkDiagram from "../components/vis/NodeLinkDiagram";
+
+return;
