+++ src/pages/MusicDetail.jsx
@@ -1,0 +1,1 @@
+
